% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CP_fit.R
\name{cp_khatri_rao}
\alias{cp_khatri_rao}
\title{Kathri-Rao matrix product}
\usage{
cp_khatri_rao(mats)
}
\arguments{
\item{mats}{A list of matrices with the same number of columns.}
}
\value{
A matrix whose \eqn{r}-th column is the Kronecker product of the
  \eqn{r}-th columns of the matrices in \code{mats}.
}
\description{
Internal helper to compute the Column-Wise Kronecker Product of a list of matrices.
All matrices must have the same number of columns. The output matrix has
the same number of columns, and each column is the Kronecker product of
the corresponding columns of the input matrices.
}
\details{
This function is used as a low-level tensor contraction primitive inside
CP tensor regression fitting routines.
}
\keyword{internal}
