% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CP_fit.R
\name{CP_normal_indep}
\alias{CP_normal_indep}
\title{Tensor-on-Tensor Regression with Independent Errors and CP Format Coefficient}
\usage{
CP_normal_indep(Yall, Xall, R, it = 100, err = 1e-08, init = NULL, retB = TRUE)
}
\arguments{
\item{Yall}{Array containing the n tensor responses along the last mode.}

\item{Xall}{Array containing the n tensor covariates along the last mode.}

\item{R}{Positive whole number indicating the CP rank.}

\item{it}{Maximum number of iterations.}

\item{err}{Relative error used to assess convergence.}

\item{init}{List containing initial values. If supplied, \code{init$CP$Ls}
and \code{init$CP$Ms} are used. Optionally, \code{init$sig2} may be used.
If \code{init = NULL}, CP factors are randomly initialized.}

\item{retB}{Return the dense tensor B of low rank?}
}
\value{
A list containing:
\code{B}, \code{CP}, \code{sig2}, \code{covs}, \code{allconv},
\code{allik}, and \code{it}.
}
\description{
Tensor-on-tensor regression
\eqn{Y_i = < X_i | B > + E_i}\cr
with independent errors
\eqn{Var(vec(E_i)) = \sigma^2 I}
and CP-formatted
\eqn{B = [[  L_1 ,.., L_l , M_1 ,..., M_p ]]}.\cr
}
\details{
This is a specialized faster version of \code{CP_normal()} for the case
where all response-mode covariance factors are identity matrices.
}
