% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datreg_sim.R
\name{datreg_sim}
\alias{datreg_sim}
\title{Data generation when X and Beta are given with types of covariance matrices}
\usage{
datreg_sim(msT, Xall, Bt, sig2t, corrs, arma_param = NULL, covars = NULL)
}
\arguments{
\item{msT}{The size of each \eqn{Y_i}, i.e., \eqn{(m_1, ..., m_p)}.}

\item{Xall}{Array containing the n tensor covariates along the last mode,
so that it is of size \eqn{h_1 \times .. \times h_l \times n}.}

\item{Bt}{The coefficient array of size \eqn{h_1 \times .. \times h_l \times m_1 \times .. \times m_p}}

\item{sig2t}{General variance \eqn{\sigma^2}}

\item{corrs}{Character vector of size p indicating the
types of covariance matrices desired for \eqn{S_1 ,.., S_p}.
Options are "AR(1)", "MA(1)", "ARMA"/"ARMA(p,q)"/"ARMA(p, q)", "EQC"  for
AR(1), MA(1), ARMA(p, q) and equivariance correlation matrices, and
"N" for general covariance with element (1,1) equal to 1.
If corrs is of size 1, then \eqn{S_1 ,.., S_p} will all have the
same correlation structure.}

\item{arma_param}{A list of size \code{length(dim(Yall))},
each of which contains the ARMA parameter orders (p, q)
for that corresponding mode.
p is the AR parameter order and q is the MA parameter order.
If some other mode has some other kind of correlation structure
and you still want to specify the ARMA orders,
you can input a list of size p with other cases as NULL.
The default ARMA order is (1, 1).}

\item{covars}{The covariance matrices \eqn{[S_1 ,.., S_p]} explicitly specified.
It would be a list of size p, where each element is a covariance matrix of
size \eqn{m_i \times m_i}.}
}
\value{
A list containing the following elements: \cr\cr
\code{Yall}: Array containing the n tensor responses along the last mode,
   so that it is of size \eqn{m_1 \times .. \times m_p \times n}.
   The last dimension must match the last dimension of \code{Xall}. \cr\cr
\code{Xall}: Array containing the n tensor covariates along the last mode,
   so that it is of size \eqn{h_1 \times .. \times h_l \times n}. \cr\cr
\code{SST}: The list of matrices \eqn{[S_1 ,.., S_p]}, \eqn{S_i} is of size \eqn{m_i \times m_i}.
}
\description{
Data generation when X and Beta are given with types of covariance matrices
}
\examples{
# Tensor-on-Tensor Regression on 6x7x8 responses and 2x3x4 covariates
set.seed(1234)
RmsT <- 6:8
Rnn <- 10
X <- array(1:prod(2:4, Rnn), dim = c(2:4, Rnn))
Bt <- array(1:prod(dim(X), RmsT), c(dim(X), RmsT))/1e5
Rcorrs <- c("ARMA", "ARMA", "N")
arma_params <- list(c(1,1), c(1, 2), NULL)
Rsig2t <- 10
dat <- datreg_sim(msT=RmsT, Xall=X, Bt=Bt, sig2t=Rsig2t,
                   corrs=Rcorrs, arma_param = arma_params)
dim(dat$Yall)
dim(dat$SST[[1]])
}
\references{
\url{https://arxiv.org/abs/2012.10249}
}
\seealso{
\code{\link{diagdat_sim}}
}
\author{
Subrata Pal, \email{SubrataluPal@gmail.com}
}
