% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OP_fit.R
\name{OP_normal_indep}
\alias{OP_normal_indep}
\title{Tensor-on-Tensor Regression with Independent Errors and OP Format Coefficient}
\usage{
OP_normal_indep(Yall, Xall, it = 100, err = 1e-07, init = NULL)
}
\arguments{
\item{Yall}{Array containing the n tensor responses along the last mode.}

\item{Xall}{Array containing the n tensor covariates along the last mode.}

\item{it}{Maximum number of iterations.}

\item{err}{Relative error used to assess convergence.}

\item{init}{List containing initial values. If supplied, \code{init$OP}
is used. Optionally, \code{init$sig2} may be used.}
}
\value{
A list containing:
\code{OP}, \code{sig2}, \code{covs}, \code{allconv}, \code{allik}, and \code{it}.
}
\description{
Tensor-on-tensor regression
\eqn{Y_i = < X_i | B > + E_i}\cr
with independent errors
\eqn{Var(vec(E_i)) = \sigma^2 I}
and OP-formatted
\eqn{B = o[[  M_1 ,..., M_p ]]}.\cr
}
\details{
This is a specialized faster version of \code{OP_normal()} for the case
where all response-mode covariance factors are identity matrices.
}
