% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{TK_predict}
\alias{TK_predict}
\title{Predict Method for Tensor-on-Tensor Regression with Tucker Decomposition}
\usage{
TK_predict(X, Tucker, retB = FALSE)
}
\arguments{
\item{X}{Tensor covariate of size \eqn{X_i} is \eqn{h_1 \times .. \times h_l}}

\item{Tucker}{List containing Tucker decomposition components:
- V: Core tensor of size \eqn{(c_1 ,.., c_l , d_1 ,.., d_p)}
- Ls: List of l matrices \eqn{L_k} of size \eqn{h_k \times c_k} 
- Ms: List of p matrices \eqn{M_k} of size \eqn{m_k \times d_k}}

\item{retB}{Logical indicating whether to return full tensor B (for small problems)}
}
\value{
Predicted tensor \eqn{Y_i} of size \eqn{m_1 \times .. \times m_p}
}
\description{
Computes the tensor product \eqn{< X_i | B >} without explicitly constructing the full B tensor.
Uses the Tucker decomposition \eqn{B = [[V; L_1,...,L_p, M_1,...,M_q]]} for efficient computation.
}
\examples{
# After fitting TK_normal model:
# fit <- TK_normal(Yall = Yall, Xall = Xall, pdims = pdims, corrs = corrs, it=2)
# Prediction for new observation new_X
prediction_fast <- TK_predict(new_X, fit$Tucker)
prediction_full <- TK_predict_full(new_X, fit$B)
norm(prediction - prediction2)
}
\references{
Llosa-Vite, C., & Maitra, R. (2022). 
  \href{https://doi.org/10.1109/TPAMI.2022.3164836}{Reduced-Rank Tensor-on-Tensor Regression and Tensor-variate Analysis of Variance}
  \emph{IEEE Transactions on Pattern Analysis and Machine Intelligence}, 45(2), 2282 - 2296.
}
\author{
Carlos Llosa-Vite, \email{llosacarlos2@gmail.com}

Subrata Pal, \email{SubrataluPal@gmail.com}

Ranjan Maitra, \email{maitra@iastate.edu}
}
