% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_UNC.R
\name{ToT_unformat}
\alias{ToT_unformat}
\title{Tensor-on-Tensor Regression where B has no format
and S has two different structures: diagonal (IndpS) and Kronecker-separable (KronS)}
\usage{
ToT_unformat(Yall, Xall, it = 100, corrs = "N", err = 1e-08)
}
\arguments{
\item{Yall}{Array containing the n tensor responses along the last mode, so that it is of size \eqn{m_1 \times .. \times m_p \times n}.
The last dimension must match the last dimension of Xall.}

\item{Xall}{Array containing the n tensor covariates along the last mode, so that it is of size \eqn{h_1 \times .. \times h_l \times n}.
The last dimension must match the last dimension of Yall.}

\item{it}{maximum number of iterations.}

\item{corrs}{Character vector of size p indicating the types of covariance matrices desired for S_1 ,.., S_p.
CHANGE THIS ONE!!}

\item{err}{relative error used to assess convergence.}
}
\value{
A list containing the following elements: \cr\cr
\code{B} -  the estimated coefficient B. \cr\cr
\code{sig2} - the estimate of \eqn{\sigma^2}. \cr\cr
\code{covs} - a list with the estimated matrices \eqn{S_1 ,.., S_p}. \cr\cr
\code{allconv} - a vector with all the convergence criteria.(????) \cr\cr
\code{allik} - a vector with all the loglikelihoods, should be monotone increasing.(????) \cr\cr
\code{it} - the number of iterations taken \cr\cr
}
\description{
Tensor-on-tensor regression
\eqn{Y_i = < X_i | B > + E_i}\cr
with kronecker-separable covariance
\eqn{Var(vec(E_i)) = \sigma^2 \otimes_k S_k}
The size of the ith tensor covariate \eqn{X_i} is \eqn{h_1 \times .. \times h_l},
The size of the ith tensor response \eqn{Y_i} is \eqn{m_1 \times .. \times m_p},
the size of the regression coefficient B is of size \eqn{h_1 \times .. \times h_l \times m_1 \times .. \times m_p},
and \eqn{i=1,...,n} .\cr
the matrix \eqn{M_k} is of size \eqn{m_k \times R},\cr
and matrix \eqn{S_k} is positive definite of size \eqn{m_k \times m_k} and is restricted to (1,1) element equal to 1.
}
\details{
Convergence is achieved when the relative difference in snorm(B) + snorm(S) is less than err.
snorm(Y) here means norm(Y)/sqrt(size of Y) \cr
Here the number of dimensions of Y and X must match. If they don't, then one can create dummy dimensions of size 1
to have them match using \code{dim(Xall) <- c(1,dim(Xall))}
}
\examples{
# Tensor-on-Tensor Regression on 6x7x8x9 responses and 3x4x5 covariates
set.seed(1234)
RmsT <- 6:9
RhsT <- 3:5
Rbt <- 100
Rnn <- 2
Rcorrs = c("AR(1)","EQC","MA(1)","N")
Rsig2t <- 20
dat <- diagdat_sim(msT=RmsT,hsT=RhsT,bt=Rbt,nn=Rnn,sig2t=Rsig2t,corrs=Rcorrs)
dim(dat$Xall) <- c(1,dim(dat$Xall))
fit <- ToT_unformat(Yall = dat$Yall,Xall = dat$Xall,corrs = Rcorrs)
par(mfrow = c(1,2))
hist(fit$B,main = "estimates in B (true in blue)")
abline(v= Rbt,col = "blue",lwd=3)
# plot(ts(fit$allik),main = "loglikelihood")
par(mfrow = c(1,1))
covars <- rbind("true" = c(Rsig2t,sapply(dat$SST,function(x)x[2,1])),
                "est" = c(fit$sig2,sapply(fit$covs,function(x)x[2,1])))
colnames(covars) <- c("sig2",paste0("S-",Rcorrs))
covars
}
\references{
\url{https://arxiv.org/abs/2012.10249}
}
\author{
Carlos Llosa-Vite, \email{llosacarlos2@gmail.com}
}
